+++ src/main/java/org/accounting/system/repositories/ResourceRepository.java
@@ -1,0 +1,9 @@
+package org.accounting.system.repositories;
+
+import jakarta.enterprise.context.ApplicationScoped;
+import org.accounting.system.entities.Resource;
+import org.accounting.system.repositories.modulators.AccessibleModulator;
+
+@ApplicationScoped
+public class ResourceRepository extends AccessibleModulator<Resource, String> {
+}