+++ src/main/java/org/accounting/system/services/ResourceService.java
@@ -1,48 +1,28 @@
 package org.accounting.system.services;
 
-import io.quarkus.cache.CacheResult;
-import io.quarkus.panache.common.Page;
 import jakarta.enterprise.context.ApplicationScoped;
 import jakarta.inject.Inject;
-import jakarta.validation.ValidationException;
-import jakarta.ws.rs.InternalServerErrorException;
-import jakarta.ws.rs.NotFoundException;
 import jakarta.ws.rs.core.UriInfo;
-import org.accounting.system.clients.ResourceClient;
-import org.accounting.system.clients.responses.eoscportal.EOSCResource;
 import org.accounting.system.dtos.pagination.PageResource;
-import org.accounting.system.dtos.resource.EoscResourceDto;
-import org.accounting.system.entities.projections.MongoQuery;
+import org.accounting.system.dtos.resource.ResourceResponse;
 import org.accounting.system.mappers.ResourceMapper;
-import org.accounting.system.util.Utility;
-import org.eclipse.microprofile.rest.client.inject.RestClient;
-import org.jboss.logging.Logger;
-
-import java.util.Collections;
-import java.util.List;
+import org.accounting.system.repositories.ResourceRepository;
 
 @ApplicationScoped
 public class ResourceService{
 
     @Inject
-    Utility utility;
-
-    @Inject
-    @RestClient
-    ResourceClient resourceClient;
-
-    @Inject
-    Logger LOG;
+    ResourceRepository resourceRepository;
 
     /**
-     * Returns the Eosc Resources from the given page.
+     * Returns the Resources from the given page.
      *
      * @param page Indicates the page number.
-     * @param size The number of Eosc Resources to be retrieved.
+     * @param size The number of Resources to be retrieved.
      * @param uriInfo The current uri.
      * @return An object represents the paginated results.
      */
-    public PageResource<EoscResourceDto> findAllEoscResources(int page, int size, UriInfo uriInfo){
+    public PageResource<ResourceResponse> findAllResources(int page, int size, UriInfo uriInfo){
 
 //        var eoscResources = getAllEoscResources();
 //
@@ -50,46 +30,13 @@ public class ResourceService{
 //
 //        var resources = partition.get(page) == null ? new ArrayList<EOSCResource>() : partition.get(page);
 
-        var pageable = new MongoQuery<EOSCResource>();
+        var resources = resourceRepository.findAllPageable(page, size);
 
-        pageable.list = Collections.emptyList();
-        pageable.index = page;
-        pageable.size = size;
-        pageable.count = 0;
-        pageable.page = Page.of(page, size);
-
-        return new PageResource<>(pageable, ResourceMapper.INSTANCE.eoscResourcesToEoscResourcesDto(pageable.list()), uriInfo);
-    }
-
-    /**
-     * Communicates with EOSC Portal to retrieve the available EOSC Resources.
-     *
-     * @return The available Eosc Resources.
-     * @throws InternalServerErrorException Error communication with Eosc Provider Portal.
-     * @throws NotFoundException Cannot found the Eosc Resources.
-     */
-    @CacheResult(cacheName = "eosc-resources")
-    public List<EOSCResource> getAllEoscResources(){
-
-        LOG.info("Retrieving the total number of EOSC Resources.");
-
-        try{
-            var total = resourceClient.getTotalNumberOfResources("all");
-
-            LOG.infof("There are %s Resources in EOSC Portal.", total.total);
-            LOG.info("Retrieving the available Resources from EOSC-Portal.");
-
-            var response = resourceClient.getAll("all", total.total);
-            return response.results;
-        } catch (ValidationException e){
-            throw new NotFoundException("Cannot found the Eosc Resources.");
-        } catch (Exception re){
-            throw new InternalServerErrorException("Failed to communicate with EOSC-Portal.");
-        }
+        return new PageResource<>(resources, ResourceMapper.INSTANCE.resourcesToResponse(resources.list()), uriInfo);
     }
 
     public boolean exist(String id){
 
-        return getAllEoscResources().stream().anyMatch(resource->resource.id.equals(id));
+       return resourceRepository.findByIdOptional(id).isPresent();
     }
 }