+++ src/main/java/org/accounting/system/services/installation/InstallationService.java
@@ -4,7 +4,6 @@ import jakarta.enterprise.context.ApplicationScoped;
 import jakarta.inject.Inject;
 import jakarta.ws.rs.ForbiddenException;
 import jakarta.ws.rs.NotFoundException;
-import jakarta.ws.rs.NotSupportedException;
 import jakarta.ws.rs.core.UriInfo;
 import org.accounting.system.dtos.acl.role.RoleAccessControlRequestDto;
 import org.accounting.system.dtos.acl.role.RoleAccessControlResponseDto;
@@ -29,7 +28,6 @@ import org.accounting.system.mappers.MetricMapper;
 import org.accounting.system.repositories.HierarchicalRelationRepository;
 import org.accounting.system.repositories.authorization.RoleRepository;
 import org.accounting.system.repositories.installation.InstallationRepository;
-import org.accounting.system.repositories.metric.MetricRepository;
 import org.accounting.system.repositories.provider.ProviderRepository;
 import org.accounting.system.services.HierarchicalRelationService;
 import org.accounting.system.services.ResourceService;
@@ -58,18 +56,13 @@ public class InstallationService implements RoleAccessControlService {
 
     @Inject
     RoleRepository roleRepository;
-
-    @Inject
-    MetricRepository metricRepository;
-
     @Inject
     HierarchicalRelationRepository hierarchicalRelationRepository;
+    @Inject
+    QueryParser queryParser;
 
     @Inject
     ResourceService resourceService;
-
-    @Inject
-    QueryParser queryParser;
 
     /**
      * Maps the {@link InstallationRequestDto} to {@link Installation}.
@@ -82,13 +75,9 @@ public class InstallationService implements RoleAccessControlService {
 
         installationRepository.exist(request.project, request.organisation, request.installation);
 
-        if(StringUtils.isNotEmpty(request.resource) ){
-            throw new NotSupportedException("Note: Some functionality is currently not available as the accounting service is yet to be integrated with the Service Catalogue that is not available at the moment in the staging environment.");
+        if(StringUtils.isNotEmpty(request.resource) && ! resourceService.exist(request.resource)){
+            throw new NotFoundException("There is no Resource with the following id: "+request.resource);
         }
-
-//        if(StringUtils.isNotEmpty(request.resource) && ! resourceService.exist(request.resource)){
-//            throw new NotFoundException("There is no Resource with the following id: "+request.resource);
-//        }
 
         var installation = installationRepository.save(request);
 
@@ -206,6 +195,7 @@ public class InstallationService implements RoleAccessControlService {
         }
 
         if(StringUtils.isNotEmpty(request.resource) && ! resourceService.exist(request.resource)){
+
             throw new NotFoundException("There is no Resource with the following id: "+request.resource);
         }