+++ CHANGELOG.md
@@ -24,9 +24,20 @@ According to [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) , the `Unr
 
 ---
 
+## 1.3.0 - 2024-05-11
+
+---
+
+### Added
+
+-   [#173](https://github.com/ARGOeu/argo-accounting/pull/173) - ACC-228 Adapt Metric schema to add support for Groups and Users.
+-   [#174](https://github.com/ARGOeu/argo-accounting/pull/174) - ACC-235 Create New Endpoint for Project Creation.
+
 ## 1.2.1 - 2024-05-22
 
 ---
+
+### Changed
 
 -   [#169](https://github.com/ARGOeu/argo-accounting/pull/169) - EOSC Resource Not Supported.